backend/pyproject.toml
[project]
name = "auctions-backend"
version = "0.1.0"
description = "Auctions backend (FastAPI + AsyncPG)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Anton Pavlov", email = "pavlov@myyhoo.com" }
]

dependencies = [
    "fastapi>=0.122.0",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0.44",
    "asyncpg>=0.31.0",
    "pydantic-settings>=2.2.1",
    "alembic>=1.12.0",
    "psycopg[binary]>=3.2.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.0",
]

[tool.uv]
managed = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"